requirements.txt
absl-py==0.13.0
astunparse==1.6.3
autopep8==1.5.7
backcall==0.2.0
blis==0.7.4
bunch==1.0.1
cachetools==4.2.2
catalogue==2.0.6
certifi==2021.5.30
charset-normalizer==2.0.6
clang==5.0
click==8.0.1
cycler==0.10.0
cymem==2.0.5
decorator==5.1.0
flatbuffers==1.12
fr-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_lg-3.1.0/fr_core_news_lg-3.1.0-py3-none-any.whl
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
idna==3.2
ipython==7.27.0
jedi==0.18.0
Jinja2==3.0.2
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
murmurhash==1.0.5
nltk==3.6.3
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.3
parso==0.8.2
pathy==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
preshed==3.0.5
prompt-toolkit==3.0.20
protobuf==3.18.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pydantic==1.8.2
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
six==1.15.0
sklearn==0.0
smart-open==5.2.1
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
thinc==8.0.10
threadpoolctl==2.2.0
toml==0.10.2
tqdm==4.62.3
traitlets==5.1.0
typer==0.4.0
typing-extensions==3.7.4.3
urllib3==1.26.6
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
yapf==0.31.0
